+++ changelog.md
@@ -7,6 +7,11 @@ and this module adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - Breaking: Split formatting role of option `logger` as separated option `messageFormatter`.
+- Breaking: `LogTags.LOG_LEVEL` renamed as `LogTags.MESSAGE_LEVEL` to distinct logging level from message level.
+- Breaking: Logging level `disabled` renamed to `off` to align with well-known log4j.
+
+### Fixed
+- Fix prefixes can be altered after set.
 
 ## [0.1.2] - 2020-09-29
 ### Fixed